+++ wtfix/apps/sessions.py
@@ -21,8 +21,6 @@ import uuid
 from asyncio import IncompleteReadError, LimitOverrunError
 from pathlib import Path
 
-from unsync import unsync
-
 from wtfix.apps.base import BaseApp
 from wtfix.conf import settings
 from wtfix.core import utils
@@ -88,7 +86,6 @@ class SessionApp(BaseApp):
                 f"{self.name}: Starting a new session with ID: {self._session_id}."
             )
 
-    @unsync
     async def initialize(self, *args, **kwargs):
         await super().initialize(*args, **kwargs)
 
@@ -123,7 +120,6 @@ class ClientSessionApp(SessionApp):
 
         self.target = target
 
-    @unsync
     async def initialize(self, *args, **kwargs):
         """
         Establish a connection to the FIX server and start listening for messages.
@@ -131,7 +127,6 @@ class ClientSessionApp(SessionApp):
         await super().initialize(*args, **kwargs)
         await self._open_connection()  # Block until connection is established
 
-    @unsync
     async def _open_connection(self):
         """
         Connect to the FIX server, obtaining StreamReader and StreamWriter instances for receiving messages
@@ -147,20 +142,18 @@ class ClientSessionApp(SessionApp):
         )
         logger.info(f"{self.name}: Connected!")
 
-    @unsync
     async def start(self, *args, **kwargs):
         """
         Start listening for messages and log on to the server.
         """
         await super().start(*args, **kwargs)
 
-        self.listen()  # Start the listener in separate task
+        asyncio.create_task(self.listen())  # Start the listener in separate task
 
         # Wait for connection to stabilize to make sure that we are listening
         # and that we do not miss any rejection messages.
         await asyncio.sleep(1)
 
-    @unsync
     async def stop(self, *args, **kwargs):
         """
         Close the writer.
@@ -172,7 +165,6 @@ class ClientSessionApp(SessionApp):
             self.writer.close()
             logger.info(f"{self.name}: Session closed!")
 
-    @unsync
     async def listen(self):
         """
         Listen for new messages that are sent by the server.
@@ -183,6 +175,8 @@ class ClientSessionApp(SessionApp):
         checksum_start = settings.SOH + utils.encode(
             f"{settings.protocol.Tag.CheckSum}="
         )
+
+        data = []
 
         while not self.writer.transport.is_closing():  # Listen forever for new messages
             try:
@@ -215,11 +209,11 @@ class ClientSessionApp(SessionApp):
                     )  # Process logout message in the pipeline as per normal
 
                 else:
-                    logger.exception(
+                    logger.error(
                         f"{self.name}: Unexpected EOF waiting for next chunk of partial data "
                         f"'{utils.decode(e.partial)}'. Initiating shutdown..."
                     )
-                    self.pipeline.stop()
+                    asyncio.create_task(self.pipeline.stop())
 
                 return
 
@@ -229,11 +223,10 @@ class ClientSessionApp(SessionApp):
                     f"{self.name}: Stream reader buffer limit exceeded! Initiating shutdown..."
                 )
 
-                self.pipeline.stop()
+                asyncio.create_task(self.pipeline.stop())
 
                 return  # Stop trying to listen for more messages.
 
-    @unsync
     async def on_send(self, message):
         """
         Writes an encoded message to the StreamWriter.