+++ wtfix/apps/utils.py
@@ -14,8 +14,6 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from unsync import unsync
-
 from wtfix.apps.base import BaseApp
 from wtfix.conf import settings
 from wtfix.message.message import FIXMessage
@@ -30,7 +28,6 @@ class OutboundLoggingApp(BaseApp):
 
     name = "outbound_logger"
 
-    @unsync
     async def on_send(self, message: FIXMessage) -> FIXMessage:
         logger.info(f" --> {message:t}")
 
@@ -44,7 +41,6 @@ class InboundLoggingApp(BaseApp):
 
     name = "inbound_logger"
 
-    @unsync
     async def on_receive(self, message: FIXMessage) -> FIXMessage:
         logger.info(f" <-- {message:t}")
 
@@ -58,6 +54,5 @@ class PipelineTerminationApp(BaseApp):
 
     name = "pipeline_termination"
 
-    @unsync
     async def on_receive(self, message: FIXMessage) -> None:
         del message